requirements.txt
Django==4.2.11
django-rest-framework==0.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
mysqlclient==2.2.4
jsonfield==3.1.0
django-rest-swagger==2.2.0